+++ packages/flutter_deck/example/lib/slides/steps_slide.dart
@@ -19,7 +19,7 @@ class StepsSlide extends FlutterDeckSlideWidget {
         );
 
   @override
-  FlutterDeckSlide build(BuildContext context) {
+  Widget build(BuildContext context) {
     return FlutterDeckSlide.split(
       leftBuilder: (context) => FlutterDeckBulletList(
         useSteps: true,