+++ packages/flutter_deck/example/lib/slides/title_slide.dart
@@ -18,7 +18,7 @@ class TitleSlide extends FlutterDeckSlideWidget {
         );
 
   @override
-  FlutterDeckSlide build(BuildContext context) {
+  Widget build(BuildContext context) {
     return FlutterDeckSlide.title(
       title: 'Welcome to flutter_deck demo! 🚀',
       subtitle: 'Use slide deck controls to navigate.',