+++ packages/flutter_deck/lib/src/flutter_deck_slide.dart
@@ -5,18 +5,13 @@ import 'package:flutter_deck/src/flutter_deck.dart';
 import 'package:flutter_deck/src/flutter_deck_speaker_info.dart';
 import 'package:flutter_deck/src/templates/templates.dart';
 import 'package:flutter_deck/src/theme/flutter_deck_theme.dart';
-import 'package:flutter_deck/src/theme/templates/flutter_deck_slide_theme.dart';
 import 'package:flutter_deck/src/widgets/internal/internal.dart';
 
-/// An abstract class that must be extended when creating a new slide for the
-/// slide deck.
+/// An abstract class that represents a slide in a slide deck.
 ///
-/// This class is used to create a new slide for the slide deck. It ensures that
-/// each slide has a defined [FlutterDeckSlideConfiguration] and a [build]
-/// method to create the slide. Diffently from the [StatelessWidget] class, the
-/// [build] method returns a [FlutterDeckSlide] instead of a [Widget].
-///
-/// Uses [FlutterDeckSlideTheme] as a base to style the slide.
+/// This class can be used to create a new slide for the slide deck. It ensures
+/// that each slide has a defined [FlutterDeckSlideConfiguration] and a [build]
+/// method to create the slide.
 ///
 /// See also:
 ///
@@ -44,7 +39,7 @@ import 'package:flutter_deck/src/widgets/internal/internal.dart';
 ///         );
 ///
 ///   @override
-///   FlutterDeckSlide build(BuildContext context) {
+///   Widget build(BuildContext context) {
 ///     return FlutterDeckSlide.blank(
 ///       theme: FlutterDeckTheme.of(context).copyWith(
 ///         slideTheme: const FlutterDeckSlideThemeData(
@@ -59,20 +54,18 @@ import 'package:flutter_deck/src/widgets/internal/internal.dart';
 ///   }
 /// }
 /// ```
-abstract class FlutterDeckSlideWidget {
+abstract class FlutterDeckSlideWidget extends StatelessWidget {
   /// Creates a new slide for the slide deck.
   ///
   /// This constructor must be called by the subclasses to create a new slide.
   /// The [configuration] argument must not be null.
   const FlutterDeckSlideWidget({
-    required this.configuration,
+    this.configuration,
+    super.key,
   });
 
   /// The configuration of the slide.
-  final FlutterDeckSlideConfiguration configuration;
-
-  /// Creates the slide.
-  FlutterDeckSlide build(BuildContext context);
+  final FlutterDeckSlideConfiguration? configuration;
 }
 
 /// The main widget for a slide in a slide deck.
@@ -82,7 +75,7 @@ abstract class FlutterDeckSlideWidget {
 /// navigation drawer.
 ///
 /// To create a new slide, use one of the named constructors.
-class FlutterDeckSlide extends StatelessWidget {
+class FlutterDeckSlide extends FlutterDeckSlideWidget {
   /// Creates a new slide.
   ///
   /// This constructor is private and should not be used directly. Instead, use
@@ -90,6 +83,7 @@ class FlutterDeckSlide extends StatelessWidget {
   const FlutterDeckSlide._({
     required WidgetBuilder builder,
     required FlutterDeckThemeData? theme,
+    super.configuration,
     super.key,
   })  : _builder = builder,
         _theme = theme;
@@ -98,6 +92,9 @@ class FlutterDeckSlide extends StatelessWidget {
   ///
   /// This constructor creates a big fact slide in a slide deck with the default
   /// header and footer.
+  ///
+  /// The [configuration] argument is optional. If not provided, the default
+  /// configuration will be used.
   ///
   /// The [title] argument must not be null. The [subtitle],
   /// [backgroundBuilder], [footerBuilder], and [headerBuilder] arguments are
@@ -109,6 +106,7 @@ class FlutterDeckSlide extends StatelessWidget {
   /// The passed [theme] will be merged with global [FlutterDeckTheme] data.
   FlutterDeckSlide.bigFact({
     required String title,
+    FlutterDeckSlideConfiguration? configuration,
     String? subtitle,
     WidgetBuilder? backgroundBuilder,
     WidgetBuilder? footerBuilder,
@@ -117,6 +115,7 @@ class FlutterDeckSlide extends StatelessWidget {
     FlutterDeckThemeData? theme,
     Key? key,
   }) : this._(
+          configuration: configuration,
           builder: (context) => FlutterDeckBigFactSlide(
             title: title,
             subtitle: subtitle,
@@ -134,18 +133,23 @@ class FlutterDeckSlide extends StatelessWidget {
   /// This constructor creates a blank slide in a slide deck with the default
   /// header and footer, and the content in-between.
   ///
+  /// The [configuration] argument is optional. If not provided, the default
+  /// configuration will be used.
+  ///
   /// The [builder] argument must not be null. The [backgroundBuilder],
   /// [footerBuilder], and [headerBuilder] arguments are optional.
   ///
   /// The passed [theme] will be merged with global [FlutterDeckTheme] data.
   FlutterDeckSlide.blank({
     required WidgetBuilder builder,
+    FlutterDeckSlideConfiguration? configuration,
     WidgetBuilder? backgroundBuilder,
     WidgetBuilder? footerBuilder,
     WidgetBuilder? headerBuilder,
     FlutterDeckThemeData? theme,
     Key? key,
   }) : this._(
+          configuration: configuration,
           builder: (context) => FlutterDeckBlankSlide(
             builder: builder,
             backgroundBuilder: backgroundBuilder,
@@ -162,14 +166,19 @@ class FlutterDeckSlide extends StatelessWidget {
   /// does not provide any default layout for the slide. It is up to the user to
   /// define it.
   ///
+  /// The [configuration] argument is optional. If not provided, the default
+  /// configuration will be used.
+  ///
   /// The [builder] argument must not be null.
   ///
   /// The passed [theme] will be merged with global [FlutterDeckTheme] data.
   const FlutterDeckSlide.custom({
     required WidgetBuilder builder,
+    FlutterDeckSlideConfiguration? configuration,
     FlutterDeckThemeData? theme,
     Key? key,
   }) : this._(
+          configuration: configuration,
           builder: builder,
           theme: theme,
           key: key,
@@ -181,6 +190,9 @@ class FlutterDeckSlide extends StatelessWidget {
   /// and footer, and the image in-between.The image can be a local asset or a
   /// network image.
   ///
+  /// The [configuration] argument is optional. If not provided, the default
+  /// configuration will be used.
+  ///
   /// The [imageBuilder] argument must not be null. The [label],
   /// [backgroundBuilder], [footerBuilder], and [headerBuilder] arguments are
   /// optional.
@@ -188,6 +200,7 @@ class FlutterDeckSlide extends StatelessWidget {
   /// The passed [theme] will be merged with global [FlutterDeckTheme] data.
   FlutterDeckSlide.image({
     required ImageBuilder imageBuilder,
+    FlutterDeckSlideConfiguration? configuration,
     String? label,
     WidgetBuilder? backgroundBuilder,
     WidgetBuilder? footerBuilder,
@@ -195,6 +208,7 @@ class FlutterDeckSlide extends StatelessWidget {
     FlutterDeckThemeData? theme,
     Key? key,
   }) : this._(
+          configuration: configuration,
           builder: (context) => FlutterDeckImageSlide(
             imageBuilder: imageBuilder,
             label: label,
@@ -211,6 +225,9 @@ class FlutterDeckSlide extends StatelessWidget {
   /// This constructor creates a quote slide in a slide deck with the default
   /// header and footer.
   ///
+  /// The [configuration] argument is optional. If not provided, the default
+  /// configuration will be used.
+  ///
   /// The [quote] argument must not be null. The [attribution],
   /// [backgroundBuilder], [footerBuilder], and [headerBuilder] arguments are
   /// optional.
@@ -221,6 +238,7 @@ class FlutterDeckSlide extends StatelessWidget {
   /// The passed [theme] will be merged with global [FlutterDeckTheme] data.
   FlutterDeckSlide.quote({
     required String quote,
+    FlutterDeckSlideConfiguration? configuration,
     String? attribution,
     WidgetBuilder? backgroundBuilder,
     WidgetBuilder? footerBuilder,
@@ -229,6 +247,7 @@ class FlutterDeckSlide extends StatelessWidget {
     FlutterDeckThemeData? theme,
     Key? key,
   }) : this._(
+          configuration: configuration,
           builder: (context) => FlutterDeckQuoteSlide(
             quote: quote,
             attribution: attribution,
@@ -248,6 +267,9 @@ class FlutterDeckSlide extends StatelessWidget {
   /// and use the [leftBuilder] and [rightBuilder] to create the content of the
   /// left and right columns.
   ///
+  /// The [configuration] argument is optional. If not provided, the default
+  /// configuration will be used.
+  ///
   /// The [leftBuilder] and [rightBuilder] arguments must not be null. The
   /// [backgroundBuilder], [footerBuilder], [headerBuilder], and [splitRatio]
   /// arguments are optional.
@@ -259,6 +281,7 @@ class FlutterDeckSlide extends StatelessWidget {
   FlutterDeckSlide.split({
     required WidgetBuilder leftBuilder,
     required WidgetBuilder rightBuilder,
+    FlutterDeckSlideConfiguration? configuration,
     WidgetBuilder? backgroundBuilder,
     WidgetBuilder? footerBuilder,
     WidgetBuilder? headerBuilder,
@@ -266,6 +289,7 @@ class FlutterDeckSlide extends StatelessWidget {
     FlutterDeckThemeData? theme,
     Key? key,
   }) : this._(
+          configuration: configuration,
           builder: (context) => FlutterDeckSplitSlide(
             leftBuilder: leftBuilder,
             rightBuilder: rightBuilder,
@@ -282,11 +306,15 @@ class FlutterDeckSlide extends StatelessWidget {
   ///
   /// This constructor creates a slide with a standard layout in a slide deck.
   ///
+  /// The [configuration] argument is optional. If not provided, the default
+  /// configuration will be used.
+  ///
   /// The [backgroundBuilder], [contentBuilder], [footerBuilder], and
   /// [headerBuilder] arguments are optional.
   ///
   /// The passed [theme] will be merged with global [FlutterDeckTheme] data.
   FlutterDeckSlide.template({
+    FlutterDeckSlideConfiguration? configuration,
     WidgetBuilder? backgroundBuilder,
     WidgetBuilder? contentBuilder,
     WidgetBuilder? footerBuilder,
@@ -294,6 +322,7 @@ class FlutterDeckSlide extends StatelessWidget {
     FlutterDeckThemeData? theme,
     Key? key,
   }) : this._(
+          configuration: configuration,
           builder: (context) => FlutterDeckSlideBase(
             backgroundBuilder: backgroundBuilder,
             contentBuilder: contentBuilder,
@@ -311,6 +340,9 @@ class FlutterDeckSlide extends StatelessWidget {
   /// the [title] and [subtitle]. Also, if the [FlutterDeckSpeakerInfo] is set,
   /// it will render the speaker info below the title and subtitle.
   ///
+  /// The [configuration] argument is optional. If not provided, the default
+  /// configuration will be used.
+  ///
   /// The [title] argument must not be null. The [subtitle],
   /// [backgroundBuilder], [footerBuilder], [headerBuilder], and
   /// [speakerInfoBuilder] arguments are optional.
@@ -318,6 +350,7 @@ class FlutterDeckSlide extends StatelessWidget {
   /// The passed [theme] will be merged with global [FlutterDeckTheme] data.
   FlutterDeckSlide.title({
     required String title,
+    FlutterDeckSlideConfiguration? configuration,
     String? subtitle,
     WidgetBuilder? backgroundBuilder,
     WidgetBuilder? footerBuilder,
@@ -326,6 +359,7 @@ class FlutterDeckSlide extends StatelessWidget {
     FlutterDeckThemeData? theme,
     Key? key,
   }) : this._(
+          configuration: configuration,
           builder: (context) => FlutterDeckTitleSlide(
             title: title,
             subtitle: subtitle,
@@ -343,7 +377,8 @@ class FlutterDeckSlide extends StatelessWidget {
   final FlutterDeckThemeData? _theme;
 
   /// Returns the current theme used by any slide.
-  /// Being used for testing purposes
+  ///
+  /// Being used for testing purposes.
   @visibleForTesting
   FlutterDeckThemeData? get theme => _theme;
 
@@ -427,6 +462,22 @@ class _SlideBody extends StatelessWidget {
             : scaffoldState.openDrawer();
       },
       child: body,
+    );
+  }
+}
+
+/// FlutterDeckSlideWidget extension methods for the [Widget] class.
+extension FlutterDeckSlideWidgetX on Widget {
+  /// Returns a [FlutterDeckSlideWidget] with the given [configuration].
+  ///
+  /// This method can be used instead of subclassing [FlutterDeckSlideWidget] to
+  /// create a new slide.
+  FlutterDeckSlideWidget withSlideConfiguration(
+    FlutterDeckSlideConfiguration configuration,
+  ) {
+    return FlutterDeckSlide.custom(
+      configuration: configuration,
+      builder: (context) => this,
     );
   }
 }