+++ packages/flutter_deck/test/src/templates/big_fact_slide_test.dart
@@ -75,7 +75,7 @@ void main() {
 }
 
 class BigFactSlide extends FlutterDeckSlideWidget {
-  const BigFactSlide()
+  const BigFactSlide({super.key})
       : super(
           configuration: const FlutterDeckSlideConfiguration(
             route: '/big-fact',
@@ -83,7 +83,7 @@ class BigFactSlide extends FlutterDeckSlideWidget {
         );
 
   @override
-  FlutterDeckSlide build(BuildContext context) {
+  Widget build(BuildContext context) {
     return FlutterDeckSlide.bigFact(
       title: '100%',
       subtitle: 'The test coverage value',